The first thing to note is that the diff is for multiple files - specifically, it looks like the diff is for the src/excel.rs file.
Looking at the diff, it looks like there is a change on line 110 - specifically, it looks like the println statement has been changed. In the old version, it looks like the println statement would print out the row number and the contents of the row. In the new version, it looks like the println statement will only print out the contents of the row.

It's important to note that this change will only affect how the data is displayed when the program is run - it won't actually change the data itself.