The diff contains changes to the main.rs and word_ipa_pair.rs files. In main.rs, a new mod, ipa_encoding_pair, was added. In word_ipa_pair.rs, the fields in the WordIpaPair struct were changed from ipa to word_ipa.
In the new ipa_encoding_pair.rs file, the IpaEncodingPair struct was created. This struct has a word field and an ipa field. The IpaEncodingPair::get_ipa_encoding_pair_list() function was also added. This function takes a selected_workbook as a parameter and returns a Vec of WordIpaPair structs.

In word_ipa_pair.rs, the WordIpaPair::get_word_ipa_pair_list() function was added. This function takes a selected_workbook and a selected_sheet as parameters and returns a Vec of WordIpaPair structs.