The first diff is for the file ipa_encoding_pair.rs. In this file, the code has been changed to replace the word_ipa variable with an arr_word_ipa vector. The code will now iterate through the ipa_encoding_pair_list and, for each pair in the list, check if the word_ipa contains the IPA. If it does, the encoding for that IPA will be added to the arr_word_ipa vector. Finally, the arr_word_ipa vector will be joined into a single string.
The second diff is for the file word_ipa_pair.rs. In this file, the code has been changed to print out the encoded_list after it has been generated. The code will now iterate through the encoded_list and print out each encoded word IPA, along with its index in the list.