The first thing to notice is that the Allophone struct has been changed to include a field for the result of the levenshtein distance calculation. This is so that we can easily print out the results of the calculation later on.
Next, the create_similarity_list function has been added. This function takes in a WordIpaPair and a list of WordIpaPairs and returns a list of Allophones. The Allophones in the list have their word and res fields populated with the data from the WordIpaPairs.

Finally, the levenshtein_distance function has been updated to take in two strings and return a float. This is so that we can get a more accurate result from the levenshtein calculation.