The first thing to note is that the diff is for multiple files. In particular, it looks like the diff is for the file excel.rs in the src directory.
The next thing to note is that the diff is for two different versions of the file. The a version is the original version, while the b version is the new version.

Finally, we can see that there are two changes in the file. The first change is on line 115, where the text i + 1 has been added. The second change is on line 116, where the text row has been replaced with Self::convert_excel_row_to_table_format(row).