The diff shows the addition of a new function, encode_word_ipa_from_excel, to the Doodle struct in the src/doodle.rs file. This function takes in a workbook and sheet name as parameters and returns a vector of Doodle objects.
The function first calls the Excel::new function to create an Excel object. It then calls the fetch_worksheet_data function on that object to get the data from the specified sheet.

Next, the function uses a map to create a vector of Doodle objects from the data. Finally, the function returns the vector of Doodle objects.