The first thing to note is that the diff is for multiple files - specifically, the src/doodle.rs file. The diff shows that the 44th line has been changed from You chose {} to You chose {}. The 7th item in the WELCOME_TEXTS array has been changed to the 6th item.
The second thing to note is that the diff also shows the index for each file. The index for the src/doodle.rs file has been changed from d42f02f..97ce919. This means that the file has been changed from one version to another.

The third and final thing to note is that the diff also shows the 100644 permission for each file. This means that the file is readable and writable by everyone.