The diff shows the changes made to two files, README.md and src/doodle.rs.
In the README.md file, a change was made to the TODO list, adding two new items and marking one as complete. An example result was also added.

In the src/doodle.rs file, a change was made to the check_encoding_cli() function. The change prints a different message to the console.

In the src/doodle.rs file, a change was made to the check_similarities_cli() function. The change prints a different message to the console.