The git diff is for the file src/cli.rs.
In the diff, lines 3-14 have been changed.

In the changed lines, the value of the constant WELCOME_TEXTS has been changed. Six strings have been replaced with seven. The new strings are the same as the old strings, except for the addition of the string 4. Check words similarities (Longest Common Subsequence).