The diff contains changes to the src/cli.rs, src/doodle.rs, and src/doodle_similarity.rs files.
In src/cli.rs, the change modifies the list of available options for the user. In particular, the change adds a new option (4. Check words similarities and Clean Data) and removes an old option (4. Check words similarities (Longest Common Subsequence)).

In src/doodle.rs, the change adds a new function, check_similarities_and_clean_cli(). This function is similar to the existing check_similarities_cli() function, but also cleans the data.

In src/doodle_similarity.rs, the change adds a new function, create_similar_doodle_list(). This function is similar to the existing create_doodle_similarity_list() function, but returns a list of similar doodles instead of a list of DoodleSimilarity objects.